move add/remove coordinates function to boardHelpers
